move coq related files to own package
