+++ requirements.txt
@@ -1,0 +1,1 @@
+pytest==7.3.1
